requirements.txt
camelot-py
pypdf
flask

transformers
modelscope
torch
